Fixes GameMapEx.cpp.
OP2Map struct was previously missing the tileHeight field, which caused many of the functions that used data based on that struct to use and/or return incorrect data. This has been corrected. Also, logTileHeight and logTileWidth were previously chars, each followed by a char[3] "padding" field. Due to the way the data is represented, it is easier to declare them as ints and remove the padding fields.

The logic in GameMapEx::GetTileEx and GameMapEx::SetTileEx was flawed in two ways, in addition to the problems caused by the incorrect OP2Map struct. First, its logic for calculating the offset of a tile in the tile data array was wrong. Second, it did not take into account that mapTileData is a pointer to the tile data array. These problems have been accounted for.
These fixes have been made by Arklon and BlackBox.

July 25 2010

git-svn-id: https://svn.outpostuniverse.org:8443/svn/outpost2/LevelsAndMods/trunk/API/HFL@706 7b3b2116-5498-11dd-abe4-fb2d8bf8a5a8